+++ game/factions/faction.py
@@ -32,6 +32,7 @@ from game.dcs.groundunittype import GroundUnitType
 from game.dcs.shipunittype import ShipUnitType
 from game.armedforces.forcegroup import ForceGroup
 from game.dcs.unittype import UnitType
+from pydcs_extensions.f16i_idf.f16i_idf import inject_F16I
 
 if TYPE_CHECKING:
     from game.theater.start_generator import ModSettings
@@ -123,6 +124,10 @@ class Faction:
     #: Note that this option cannot be set per-side. If either faction needs it,
     #: both will use it.
     unrestricted_satnav: bool = False
+
+    # Store mod settings so mod properties can be injected again on game load,
+    # in case mods like CJS F/A-18E/F/G or IDF F-16I are selected by the player
+    mod_settings: Optional[ModSettings] = field(default=None)
 
     def has_access_to_dcs_type(self, unit_type: Type[DcsUnitType]) -> bool:
         # Vehicle and Ship Units
@@ -294,7 +299,21 @@ class Faction:
             if unit.unit_class is unit_class:
                 yield unit
 
-    def apply_mod_settings(self, mod_settings: ModSettings) -> None:
+    def apply_mod_settings(self, mod_settings: Optional[ModSettings] = None) -> None:
+        if mod_settings is None:
+            if self.mod_settings is None:
+                # No mod settings were provided and none were saved for this faction
+                # so stop here
+                return
+            elif self.mod_settings is not None:
+                # Saved mod settings were found for this faction,
+                # so load them for use
+                mod_settings = self.mod_settings
+        else:
+            # Update the mod settings of this faction
+            # so the settings can be applied again on load, if needed
+            self.mod_settings = mod_settings
+
         # aircraft
         if not mod_settings.a4_skyhawk:
             self.remove_aircraft("A-4E-C")
@@ -307,6 +326,14 @@ class Faction:
             self.remove_aircraft("VSN_F4B")
         if not mod_settings.f15d_baz:
             self.remove_aircraft("F-15D")
+        if not mod_settings.f_16_idf:
+            self.remove_aircraft("F-16I")
+            self.remove_aircraft("F_16D_52")
+            self.remove_aircraft("F_16D_50")
+            self.remove_aircraft("F_16D_50_NS")
+            self.remove_aircraft("F_16D_52_NS")
+        else:
+            inject_F16I()
         if not mod_settings.f22_raptor:
             self.remove_aircraft("F-22A")
         if not mod_settings.f100_supersabre: