+++ game/theater/start_generator.py
@@ -54,6 +54,7 @@ class ModSettings:
     a4_skyhawk: bool = False
     f4b_phantom: bool = False
     f15d_baz: bool = False
+    f_16_idf: bool = False
     f22_raptor: bool = False
     f100_supersabre: bool = False
     f104_starfighter: bool = False